remove snafu dep and global error enum
