+++ builder/Cargo.toml
@@ -40,7 +40,6 @@ rand = "0.8.5"
 sequencer = { path = "../sequencer", features = ["testing"] }
 sequencer-utils = { path = "../utils" }
 serde = { workspace = true }
-snafu = { workspace = true }
 surf = "2.3.1"
 surf-disco = { workspace = true }
 tide-disco = { workspace = true }
