+++ hotshot-state-prover/src/bin/state-prover.rs
@@ -1,8 +1,7 @@
-use std::{str::FromStr as _, time::Duration};
+use std::time::Duration;
 
 use clap::Parser;
-use cld::ClDuration;
-use espresso_types::SeqTypes;
+use espresso_types::{parse_duration, SeqTypes};
 use ethers::{
     providers::{Http, Middleware, Provider},
     signers::{coins_bip39::English, MnemonicBuilder, Signer},
@@ -12,7 +11,6 @@ use hotshot_stake_table::config::STAKE_TABLE_CAPACITY;
 use hotshot_state_prover::service::{run_prover_once, run_prover_service, StateProverConfig};
 use hotshot_types::traits::node_implementation::NodeType;
 use sequencer_utils::logging;
-use snafu::Snafu;
 use url::Url;
 use vbs::version::StaticVersionType;
 
@@ -83,19 +81,6 @@ struct Args {
 
     #[clap(flatten)]
     logging: logging::Config,
-}
-
-#[derive(Clone, Debug, Snafu)]
-pub struct ParseDurationError {
-    reason: String,
-}
-
-fn parse_duration(s: &str) -> Result<Duration, ParseDurationError> {
-    ClDuration::from_str(s)
-        .map(Duration::from)
-        .map_err(|err| ParseDurationError {
-            reason: err.to_string(),
-        })
 }
 
 #[async_std::main]
