+++ sequencer/src/bin/orchestrator.rs
@@ -2,13 +2,12 @@ use std::{num::NonZeroUsize, time::Duration};
 
 use clap::Parser;
 use derive_more::From;
-use espresso_types::PubKey;
+use espresso_types::{parse_duration, PubKey, Ratio};
 use ethers::utils::hex::{self, FromHexError};
 use hotshot_orchestrator::{
     config::{Libp2pConfig, NetworkConfig},
     run_orchestrator,
 };
-use sequencer::options::{parse_duration, Ratio};
 use sequencer_utils::logging;
 use snafu::Snafu;
 use url::Url;
