+++ types/Cargo.toml
@@ -43,7 +43,6 @@ sequencer-utils = { path = "../utils" }
 serde = { workspace = true }
 serde_json = { workspace = true }
 sha2 = "0.10"                                                        # TODO temporary, used only for VID, should be set in hotshot
-snafu = { workspace = true }
 static_assertions = { workspace = true }
 surf-disco = { workspace = true }
 tagged-base64 = { workspace = true }
