+++ types/src/v0/impls/mod.rs
@@ -12,6 +12,6 @@ mod state;
 mod transaction;
 
 pub use fee_info::FeeError;
-pub use header::ProposalValidationError;
 pub use instance_state::{mock, NodeState};
+pub use state::ProposalValidationError;
 pub use state::{validate_proposal, BuilderValidationError, StateValidationError, ValidatedState};
